+++ src/vector_errors.h
@@ -2,7 +2,7 @@
 #ifndef VECTOR_ERROR_H
 #define	VECTOR_ERROR_H
 
-int GenBERErrorVector(int *, int , int , float );
+int GenBERErrorVector(int *, int , int , double );
 int CompareVectors(uint8_t *, uint8_t *, int );
 
 #endif
